requirements.txt
asgiref==3.9.1
certifi==2025.7.9
charset-normalizer==3.4.2
Django==5.2.4
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
gunicorn==21.2.0
django-cors-headers==4.3.1
python-decouple
django-storages
Pillow
boto3
drf-yasg>=1.21.5